Limite la taille du diff à master, extrait le refactor dans un fichier dédié dans une autre PR refacto pur
