+++ tools/check-links-validity.ts
@@ -1,7 +1,6 @@
-import {
-  getRequiredAdditionsAndTouchWarningsToKeep,
-  getBenefitData,
-} from "../lib/benefits/link-validity.js"
+import Benefits from "../data/all.js"
+import config from "../backend/config/index.js"
+import { getRequiredAdditionsAndTouchWarningsToKeep } from "../lib/benefits/link-validity.js"
 import { GristResponse, GristUpdate } from "../lib/types/link-validity.js"
 
 import axios from "axios"
@@ -10,6 +9,32 @@ import Bluebird from "bluebird"
 
 // Avoid some errors due to bad tls management
 const httpsAgent = new https.Agent({ rejectUnauthorized: false })
+
+const customBenefitsFiles = [
+  {
+    pattern: /-fsl-eligibilite$/,
+    file: `${config.github.repository_url}/blob/master/data/benefits/dynamic/fsl.ts`,
+  },
+  {
+    pattern: /-apa-eligibilite$/,
+    file: `${config.github.repository_url}/blob/master/data/benefits/dynamic/apa.ts`,
+  },
+  {
+    pattern: /^aidesvelo_/,
+    file: `https://github.com/mquandalle/mesaidesvelo/blob/master/src/aides.yaml`,
+  },
+]
+
+function setEditLink(benefit) {
+  for (const category of customBenefitsFiles) {
+    if (benefit.id.match(category.pattern)) {
+      return category.file
+    }
+  }
+  return ["openfisca", "javascript"].includes(benefit.source)
+    ? `https://contribuer-aides-jeunes.netlify.app/admin/#/collections/benefits_${benefit.source}/entries/${benefit.id}`
+    : undefined
+}
 
 async function checkURL(benefit) {
   const results = await Bluebird.map(benefit.links, fetchStatus)
@@ -54,6 +79,70 @@ async function getHTTPStatus(link) {
 
 function sleep(ms) {
   return new Promise((resolve) => setTimeout(resolve, ms))
+}
+
+async function getPriorityStats() {
+  const stats = await axios
+    .get("https://aides-jeunes-stats-recorder.osc-fr1.scalingo.io/statistics")
+    .then((r) => r.data)
+  const statTotal = stats.map((v) => {
+    const p = v.events?.showDetails || {}
+    const totals = Object.keys(p)
+    return {
+      benefit: v.benefit,
+      count: totals.reduce((at, t) => {
+        const indexes = Object.keys(p[t])
+        return (
+          at +
+          indexes.reduce((ai, i) => {
+            return ai + p[t][i]
+          }, 0)
+        )
+      }, 0),
+    }
+  })
+
+  const statByBenefitId = statTotal.reduce((a, v) => {
+    a[v.benefit] = v.count
+    return a
+  }, {})
+
+  return statByBenefitId
+}
+
+async function getBenefitData(noPriority: boolean) {
+  const priorityMap = noPriority ? {} : await getPriorityStats()
+
+  const data = Benefits.all.map((benefit) => {
+    const linkMap = ["link", "instructions", "form", "teleservice"]
+      .filter(
+        (linkType) => benefit[linkType] && typeof benefit[linkType] === "string"
+      )
+      .reduce((a, linkType) => {
+        const link = benefit[linkType]
+        a[link] = a[link] || { link, types: [] }
+        a[link].types.push(linkType)
+
+        return a
+      }, {})
+
+    const links = Object.values(linkMap).map((v: any) => {
+      return {
+        link: v.link,
+        type: v.types.join(" / "),
+      }
+    })
+
+    return {
+      id: benefit.id,
+      label: benefit.label,
+      institution: benefit.institution.label,
+      priority: priorityMap[benefit.id] || 0,
+      links,
+      editLink: setEditLink(benefit),
+    }
+  })
+  return data.sort((a, b) => +(a.priority - b.priority))
 }
 
 const docId = process.env.GRIST_DOC_ID
